+++ source/databricks/package/integration_events_persister.py
@@ -14,10 +14,15 @@
 
 import sys
 
-sys.path.append(r"/workspaces/opengeh-wholesale/source/databricks")
-
 from pyspark.sql import DataFrame
-from pyspark.sql.functions import year, month, dayofmonth, col, from_json, current_timestamp
+from pyspark.sql.functions import (
+    year,
+    month,
+    dayofmonth,
+    col,
+    from_json,
+    current_timestamp,
+)
 
 
 # integration_events_persister
@@ -28,8 +33,7 @@ def integration_events_persister(
 ):
 
     events = (
-        streamingDf
-        .withColumn("storedTime", current_timestamp())
+        streamingDf.withColumn("storedTime", current_timestamp())
         .withColumn("year", year(col("storedTime")))
         .withColumn("month", month(col("storedTime")))
         .withColumn("day", dayofmonth(col("storedTime")))
