+++ source/databricks/package/integration_events_persister_streaming.py
@@ -12,11 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import sys
-
-sys.path.append(r"/workspaces/opengeh-wholesale/source/databricks")
-sys.path.append(r"/opt/conda/lib/python3.8/site-packages")
-
 from package import integration_events_persister, initialize_spark
 import configargparse
