+++ source/databricks/tests/conftest.py
@@ -16,6 +16,7 @@ By having a conftest.py in this directory, we are able to add all packages
 defined in the geh_stream directory in our tests.
 """
 
+import os
 import pytest
 from pyspark import SparkConf
 from pyspark.sql import SparkSession
@@ -35,3 +36,40 @@ def spark():
     )
 
     return SparkSession.builder.config(conf=spark_conf).getOrCreate()
+
+
+@pytest.fixture(scope="session")
+def file_path_finder():
+    """
+    Returns the path of the file.
+    Please note that this only works if current folder haven't been changed prior using `os.chdir()`.
+    The correctness also relies on the prerequisite that this function is actually located in a
+    file located directly in the integration tests folder.
+    """
+
+    def finder(file):
+        return os.path.dirname(os.path.normpath(file))
+
+    return finder
+
+
+@pytest.fixture(scope="session")
+def source_path(file_path_finder) -> str:
+    """
+    Returns the <repo-root>/source folder path.
+    Please note that this only works if current folder haven't been changed prior using `os.chdir()`.
+    The correctness also relies on the prerequisite that this function is actually located in a
+    file located directly in the integration tests folder.
+    """
+    return file_path_finder(f"{__file__}/../..")
+
+
+@pytest.fixture(scope="session")
+def databricks_path(source_path) -> str:
+    """
+    Returns the source/databricks folder path.
+    Please note that this only works if current folder haven't been changed prior using `os.chdir()`.
+    The correctness also relies on the prerequisite that this function is actually located in a
+    file located directly in the integration tests folder.
+    """
+    return f"{source_path}/databricks"
